requirements.txt
Django==1.10.2
django-crispy-forms==1.6.0
django-registration-redux==1.4
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
djangorestframework==3.5.3
